feat: add two new custom controls for select, multi select and file upload
